requirements.txt
annotated-types==0.7.0
bidict==0.23.1
blinker==1.9.0
cachelib==0.13.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
dash==3.0.4
dash-bootstrap-components==2.0.2
dash-extensions==2.0.4
dash_ag_grid==31.3.1
dataclass-wizard==0.35.0
dnspython==2.7.0
EditorConfig==0.17.0
eventlet==0.39.1
filelock==3.18.0
Flask==3.0.3
Flask-Caching==2.3.1
Flask-SocketIO==5.5.1
fonttools==4.57.0
fsspec==2025.3.2
greenlet==3.2.1
h11==0.14.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.0
jsbeautifier==1.15.4
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
more-itertools==10.7.0
mpmath==1.3.0
narwhals==1.37.1
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.1.1
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plotly==6.0.1
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic==2.11.4
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-engineio==4.12.0
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
retrying==1.3.4
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==78.1.0
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.40
sympy==1.13.1
threadpoolctl==3.6.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
ultralytics==8.3.102
ultralytics-thop==2.0.14
urllib3==2.3.0
websocket-client==1.8.0
Werkzeug==3.0.6
wsproto==1.2.0
zipp==3.21.0
